lab 3 refactored

+++ include/iofunc.h
@@ -27,6 +27,6 @@ void writeBinData(const std::string, const std::vector<float> &);
 
 void readRawData(const std::string in_fname, std::vector<uint8_t> &raw_data);
 
-void convertRaw(const std::vector<uint8_t> raw_data, std::vector<double> &iq_data);
+void convertRaw(const std::vector<uint8_t> raw_data, std::vector<float> &iq_data);
 
 #endif // DY4_IOFUNC_H
